requirements.txt
# If you need to access a server with the Kerberos SASL authentication mechanism you must install the gssapi or the winkerberos package
ldap3[gssapi]==2.9
#ldap3[winkerberos]==2.9
uvicorn[standard]==0.23.2
alibabacloud-dysmsapi20170525==4.5.0
alibabacloud-dysmsapi20180501==1.0.10
aliyun-python-sdk-core==2.16.0
dingtalk-sdk==1.3.8
django-redis==4.12.1
eventlet==0.40.2
gunicorn==21.2.0
httptools==0.7.1
httpx==0.25.1
psutil==7.0.0
pycryptodome==3.10.1
python-dateutil==2.9.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
setuptools==68.2.2
tencentcloud-sdk-python==3.1.34
watchfiles==1.1.1
websockets==16.0
